EventHive/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Bcrypt==1.0.1
Flask-Mail==0.9.1
Flask-Login==0.6.2
razorpay==1.4.4
qrcode[pil]==7.4.2
reportlab==4.0.4
Pillow==10.0.1